automation/finops/requirements.txt
google-cloud-bigquery==3.27.0
prometheus-client==0.21.1
pytest==8.3.5
